EPMLABSBRN-306 move csv records processing methods to processors classes <eom>
