+++ src/main/kotlin/com/epam/brn/upload/csv/processor/GroupRecordProcessor.kt
@@ -1,12 +1,14 @@
 package com.epam.brn.upload.csv.processor
 
 import com.epam.brn.model.ExerciseGroup
+import com.epam.brn.repo.ExerciseGroupRepository
 import com.epam.brn.upload.csv.record.GroupRecord
 import org.springframework.stereotype.Component
 
 @Component
-class GroupRecordProcessor {
+class GroupRecordProcessor(private val groupRepository: ExerciseGroupRepository) {
 
-    fun convert(source: GroupRecord) =
-        ExerciseGroup(name = source.name, description = source.description, id = source.groupId)
+    fun process(records: MutableList<GroupRecord>): Iterable<ExerciseGroup> {
+        return groupRepository.saveAll(records.map { ExerciseGroup(it) })
+    }
 }
