+++ src/main/kotlin/com/epam/brn/upload/csv/processor/SeriesGenericRecordProcessor.kt
@@ -1,23 +1,21 @@
 package com.epam.brn.upload.csv.processor
 
 import com.epam.brn.model.Series
+import com.epam.brn.repo.SeriesRepository
 import com.epam.brn.service.ExerciseGroupsService
 import com.epam.brn.upload.csv.record.SeriesGenericRecord
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
 @Component
-class SeriesGenericRecordProcessor {
+class SeriesGenericRecordProcessor(
+    private val seriesRepository: SeriesRepository,
+    private val exerciseGroupsService: ExerciseGroupsService
+) {
 
-    @Autowired
-    lateinit var exerciseGroupsService: ExerciseGroupsService
+    fun process(records: MutableList<SeriesGenericRecord>): Iterable<Series> {
+        val result = records
+            .map { Series(it, exerciseGroupsService.findGroupById(it.groupId)) }
 
-    fun convert(source: SeriesGenericRecord): Series {
-        return Series(
-            name = source.name,
-            description = source.description,
-            exerciseGroup = exerciseGroupsService.findGroupById(source.groupId),
-            id = source.seriesId
-        )
+        return seriesRepository.saveAll(result)
     }
 }
