+++ src/main/kotlin/com/epam/brn/upload/csv/processor/SeriesThreeExerciseRecordProcessor.kt
@@ -1,0 +1,93 @@
+package com.epam.brn.upload.csv.processor
+
+import com.epam.brn.constant.ExerciseTypeEnum
+import com.epam.brn.constant.WordTypeEnum
+import com.epam.brn.model.Exercise
+import com.epam.brn.model.Resource
+import com.epam.brn.model.Task
+import com.epam.brn.service.ExerciseService
+import com.epam.brn.service.SeriesService
+import org.springframework.stereotype.Component
+
+@Component
+class SeriesThreeExerciseRecordProcessor(
+    private val seriesService: SeriesService,
+    private val exerciseService: ExerciseService
+) {
+
+    fun process(records: MutableList<Map<String, Any>>): List<Exercise> {
+        return exerciseService.save(createExercises())
+    }
+
+    private fun createExercises(): List<Exercise> {
+        val exercise = createExercise()
+        exercise.addTask(createTask())
+        return listOf(exercise)
+    }
+
+    private fun createExercise(): Exercise {
+        return Exercise(
+            series = seriesService.findSeriesForId(3L),
+            name = "Распознование предложений из 2 слов",
+            description = "Распознование предложений из 2 слов",
+            template = "<OBJECT OBJECT_ACTION>",
+            exerciseType = ExerciseTypeEnum.SENTENCE.toString(),
+            level = 1
+        )
+    }
+
+    private fun createTask(): Task {
+        val resource1 = Resource(
+            word = "девочкаTest",
+            wordType = WordTypeEnum.OBJECT.toString(),
+            audioFileUrl = "series2/девочка.mp3",
+            pictureFileUrl = "pictures/withWord/девочка.jpg"
+        )
+        val resource2 = Resource(
+            word = "дедушкаTest",
+            wordType = WordTypeEnum.OBJECT.toString(),
+            audioFileUrl = "series2/дедушка.mp3",
+            pictureFileUrl = "pictures/withWord/дедушка.jpg"
+        )
+        val resource3 = Resource(
+            word = "бабушкаTest",
+            wordType = WordTypeEnum.OBJECT.toString(),
+            audioFileUrl = "series2/бабушка.mp3",
+            pictureFileUrl = "pictures/withWord/бабушка.jpg"
+        )
+        val resource4 = Resource(
+            word = "бросаетTest",
+            wordType = WordTypeEnum.OBJECT_ACTION.toString(),
+            audioFileUrl = "series2/бросает.mp3",
+            pictureFileUrl = "pictures/withWord/бросает.jpg"
+        )
+        val resource5 = Resource(
+            word = "читаетTest",
+            wordType = WordTypeEnum.OBJECT_ACTION.toString(),
+            audioFileUrl = "series2/читает.mp3",
+            pictureFileUrl = "pictures/withWord/читает.jpg"
+        )
+        val resource6 = Resource(
+            word = "рисуетTest",
+            wordType = WordTypeEnum.OBJECT_ACTION.toString(),
+            audioFileUrl = "series2/рисует.mp3",
+            pictureFileUrl = "pictures/withWord/рисует.jpg"
+        )
+
+        val answerOptions =
+            mutableSetOf(resource1, resource2, resource3, resource4, resource5, resource6)
+
+        val correctAnswer = Resource(
+            word = "девочка рисует",
+            wordType = WordTypeEnum.SENTENCE.toString(),
+            audioFileUrl = "series3/девочка_рисует.mp3"
+        )
+
+        return Task(
+            serialNumber = 2,
+            answerOptions = answerOptions,
+            correctAnswer = correctAnswer,
+            answerParts = mutableMapOf(1 to resource1, 2 to resource6)
+        )
+    }
+}
